mai-advisor-mcp/huggingface_space_deploy/requirements.txt
# Core dependencies for MAI Advisor MCP
gradio==5.49.1
pydantic>=2.0.0

# Optional: AI Research Agent (comment out for basic demo)
# gradio==5.49.1
langchain>=0.3.13
langchain-openai>=0.2.10
tavily-python>=0.5.0
openai>=1.54.0
mcp>=1.0.0
# anthropic>=0.39.0
# langchain>=0.3.0
# langchain-anthropic>=0.3.0
